+++ cocos2d-ui/Platform/CCPlatformTextField.m
@@ -9,7 +9,7 @@
 #import "CCPlatformTextField.h"
 
 @implementation CCPlatformTextField
-- (void) positionInControl:(CCControl *)control padding:(float)padding {
+- (void) positionInControl:(CCControl *)control padding:(CGFloat)padding {
     
 }
 - (void) onEnterTransitionDidFinish {
